pyproject.toml
[project]
name = "dev-new-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.0",
    "google-generativeai>=0.8.3",
    "pydantic-ai>=0.0.19",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
    "ruff>=0.9.2",
]
